Move getAddressFromScript to encoding.js
